Tried to refactor getcalories out of homepage
