+++ src/main/java/use_case/GetCalories/GetCaloriesOutputBoundary.java
@@ -4,4 +4,6 @@ public interface GetCaloriesOutputBoundary {
     void prepareSuccessView(GetCaloriesOutputData getCaloriesOutputData);
     void prepareFailView(String error);
     void prepareHomeView();
+
+    void prepareGetCaloriesView(String username);
 }
